requirements.txt
google-cloud-pubsub==2.7.0
requests==2.26.0